refactor(tasks): finish locomotion common rewards [skip ci]
