--- src/unilab/envs/locomotion/common/__init__.py
@@ -1,5 +1,0 @@
-from .rewards import RewardContext
-
-__all__ = [
-    "RewardContext",
-]
